feedback_analyzer1/requirements.txt
streamlit
pandas
python-dotenv
matplotlib
pyngrok
scikit-learn